Finally adding the factory method.
The previous example only got us partway to the parallel
creator/product class structure of the factory pattern.

Splitting responsibility of creating the pizzas and which pizzas to
create requires changing the pizzastore to an abstract class with a
factory method createPizza.

With this in place we can now add new stores that create pizzas without
updating the pizza classes; and add new pizza types without having to
modify the stores. The stores can each contain their own subset of
supported pizza types.